feat(Swedbank Fallback): add retry filter for handshake excepttion

+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/se/banks/swedbank/fallback/SwedbankFallbackConstants.java
@@ -57,6 +57,6 @@ public class SwedbankFallbackConstants {
 
     public static class Filters {
         public static final int NUMBER_OF_RETRIES = 5;
-        public static final long MS_TO_WAIT = 4000;
+        public static final int MS_TO_WAIT = 4000;
     }
 }
